+++ gen/com/intellij/idea/plugin/hybris/impex/psi/ImpexUserRightsEnd.java
@@ -25,6 +25,6 @@ import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
 
-public interface ImpexUserRightsEnd extends ImpexUserRightsAwarePsiElement {
+public interface ImpexUserRightsEnd extends PsiElement {
 
 }