+++ gen/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexUserRightsFirstValueGroupImpl.java
@@ -28,9 +28,10 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.util.PsiTreeUtil;
 import static com.intellij.idea.plugin.hybris.impex.psi.ImpexTypes.*;
+import com.intellij.extapi.psi.ASTWrapperPsiElement;
 import com.intellij.idea.plugin.hybris.impex.psi.*;
 
-public class ImpexUserRightsFirstValueGroupImpl extends ImpexUserRightsAwarePsiElementImpl implements ImpexUserRightsFirstValueGroup {
+public class ImpexUserRightsFirstValueGroupImpl extends ASTWrapperPsiElement implements ImpexUserRightsFirstValueGroup {
 
   public ImpexUserRightsFirstValueGroupImpl(@NotNull ASTNode node) {
     super(node);