+++ gen/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexUserRightsHeaderParameterImpl.java
@@ -28,10 +28,11 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.util.PsiTreeUtil;
 import static com.intellij.idea.plugin.hybris.impex.psi.ImpexTypes.*;
+import com.intellij.extapi.psi.ASTWrapperPsiElement;
 import com.intellij.idea.plugin.hybris.impex.psi.*;
 import java.util.Collection;
 
-public class ImpexUserRightsHeaderParameterImpl extends ImpexUserRightsAwarePsiElementImpl implements ImpexUserRightsHeaderParameter {
+public class ImpexUserRightsHeaderParameterImpl extends ASTWrapperPsiElement implements ImpexUserRightsHeaderParameter {
 
   public ImpexUserRightsHeaderParameterImpl(@NotNull ASTNode node) {
     super(node);