+++ gen/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexUserRightsValueGroupImpl.java
@@ -28,9 +28,10 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.util.PsiTreeUtil;
 import static com.intellij.idea.plugin.hybris.impex.psi.ImpexTypes.*;
+import com.intellij.extapi.psi.ASTWrapperPsiElement;
 import com.intellij.idea.plugin.hybris.impex.psi.*;
 
-public class ImpexUserRightsValueGroupImpl extends ImpexUserRightsAwarePsiElementImpl implements ImpexUserRightsValueGroup {
+public class ImpexUserRightsValueGroupImpl extends ASTWrapperPsiElement implements ImpexUserRightsValueGroup {
 
   public ImpexUserRightsValueGroupImpl(@NotNull ASTNode node) {
     super(node);