+++ gen/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexUserRightsValueImpl.java
@@ -30,7 +30,7 @@ import com.intellij.psi.util.PsiTreeUtil;
 import static com.intellij.idea.plugin.hybris.impex.psi.ImpexTypes.*;
 import com.intellij.idea.plugin.hybris.impex.psi.*;
 
-public class ImpexUserRightsValueImpl extends ImpexUserRightsAwarePsiElementImpl implements ImpexUserRightsValue {
+public class ImpexUserRightsValueImpl extends ImpexUserRightsValueMixin implements ImpexUserRightsValue {
 
   public ImpexUserRightsValueImpl(@NotNull ASTNode node) {
     super(node);